feat: Refactor project structure and update configurations for Node.js
