+++ app/portal/.dockerignore
@@ -1,6 +1,4 @@
-no.DS_Store
-.env
-/node_modules/
-
-# React Router
-/.react-router/
+.react-router
+build
+node_modules
+README.md
