+++ app/portal/app/routes/chat.$sessionId.tsx
@@ -1,20 +1,40 @@
-import React, { useState, useRef, useEffect } from 'react';
-import { Send, Bot, User, Code, Database, Globe, Zap, ExternalLink, Copy, Download, MessageCircle, Shield } from 'lucide-react';
-import { Form, useLoaderData, useActionData, useFetcher, Link, redirect } from 'react-router';
-import type { Route } from './+types/chat.$sessionId';
+import React, { useState, useRef, useEffect } from "react";
+import {
+  Send,
+  Bot,
+  User,
+  Code,
+  Database,
+  Globe,
+  Zap,
+  ExternalLink,
+  Copy,
+  Download,
+  MessageCircle,
+  Shield,
+} from "lucide-react";
+import {
+  Form,
+  useLoaderData,
+  useActionData,
+  useFetcher,
+  Link,
+  redirect,
+} from "react-router";
+import type { Route } from "./+types/chat.$sessionId";
 
 interface Message {
   id: string;
-  type: 'user' | 'agent' | 'system';
+  type: "user" | "agent" | "system";
   content: string;
   timestamp: string;
   tools?: string[];
-  status?: 'processing' | 'completed' | 'error';
+  status?: "processing" | "completed" | "error";
   artifacts?: DeploymentArtifact[];
 }
 
 interface DeploymentArtifact {
-  type: 'url' | 'config' | 'code' | 'documentation';
+  type: "url" | "config" | "code" | "documentation";
   title: string;
   content: string;
   description?: string;
@@ -25,7 +45,7 @@ interface ConsentRequest {
   workloadId: string;
   requiredScopes: string[];
   reason: string;
-  status: 'pending' | 'approved' | 'denied';
+  status: "pending" | "approved" | "denied";
 }
 
 // Store chat state in memory (in real app, this would be in database)
@@ -35,43 +55,48 @@ let pendingConsentRequests: ConsentRequest[] = [];
 export function meta({ params }: Route.MetaArgs) {
   return [
     { title: `AI Chat - Session ${params.sessionId}` },
-    { name: "description", content: "Chat with AI agent with consent-aware tool access" },
+    {
+      name: "description",
+      content: "Chat with AI agent with consent-aware tool access",
+    },
   ];
 }
 
 export function loader({ params }: Route.LoaderArgs) {
   const { sessionId } = params;
-  
+
   // Initialize session if it doesn't exist
   if (!chatSessions[sessionId]) {
     chatSessions[sessionId] = [
       {
-        id: '1',
-        type: 'system',
+        id: "1",
+        type: "system",
         content: `Connected to session ${sessionId}. I can help you with file operations, data analysis, and system tasks - but I'll need your consent for specific tool access.`,
         timestamp: new Date().toISOString(),
-        status: 'completed'
-      }
+        status: "completed",
+      },
     ];
   }
-  
+
   return {
     messages: chatSessions[sessionId],
     sessionId,
-    pendingRequests: pendingConsentRequests.filter(r => r.workloadId === sessionId)
+    pendingRequests: pendingConsentRequests.filter(
+      (r) => r.workloadId === sessionId
+    ),
   };
 }
 
 export async function action({ request, params }: Route.ActionArgs) {
   const { sessionId } = params;
   const formData = await request.formData();
-  const intent = formData.get('intent') as string;
+  const intent = formData.get("intent") as string;
 
-  if (intent === 'send-message') {
-    const message = formData.get('message') as string;
+  if (intent === "send-message") {
+    const message = formData.get("message") as string;
 
     if (!message.trim()) {
-      return { success: false, error: 'Message cannot be empty' };
+      return { success: false, error: "Message cannot be empty" };
     }
 
     // Initialize session if it doesn't exist
@@ -82,36 +107,38 @@ export async function action({ request, params }: Route.ActionArgs) {
     // Add user message
     const userMessage: Message = {
       id: Date.now().toString(),
-      type: 'user',
+      type: "user",
       content: message,
       timestamp: new Date().toISOString(),
-      status: 'completed'
+      status: "completed",
     };
 
     chatSessions[sessionId].push(userMessage);
 
     // Check if message requires consent
-    const requiresWriteAccess = message.toLowerCase().includes('create') || 
-                               message.toLowerCase().includes('write') ||
-                               message.toLowerCase().includes('modify') ||
-                               message.toLowerCase().includes('update');
-    
-    const requiresExportAccess = message.toLowerCase().includes('export') || 
-                                message.toLowerCase().includes('download') ||
-                                message.toLowerCase().includes('backup');
+    const requiresWriteAccess =
+      message.toLowerCase().includes("create") ||
+      message.toLowerCase().includes("write") ||
+      message.toLowerCase().includes("modify") ||
+      message.toLowerCase().includes("update");
+
+    const requiresExportAccess =
+      message.toLowerCase().includes("export") ||
+      message.toLowerCase().includes("download") ||
+      message.toLowerCase().includes("backup");
 
     if (requiresWriteAccess || requiresExportAccess) {
       const requiredScopes = [];
-      if (requiresWriteAccess) requiredScopes.push('tools:write');
-      if (requiresExportAccess) requiredScopes.push('data:export');
+      if (requiresWriteAccess) requiredScopes.push("tools:write");
+      if (requiresExportAccess) requiredScopes.push("data:export");
 
       // Create consent request
       const consentRequest: ConsentRequest = {
         id: `consent-${Date.now()}`,
         workloadId: sessionId,
         requiredScopes,
-        reason: `Agent needs ${requiredScopes.join(', ')} permissions to complete: "${message}"`,
-        status: 'pending'
+        reason: `Agent needs ${requiredScopes.join(", ")} permissions to complete: "${message}"`,
+        status: "pending",
       };
 
       pendingConsentRequests.push(consentRequest);
@@ -119,22 +146,23 @@ export async function action({ request, params }: Route.ActionArgs) {
       // Add agent response requesting consent
       const agentMessage: Message = {
         id: (Date.now() + 1).toString(),
-        type: 'agent',
-        content: `I need your consent to access the following scopes to complete your request:\n\n` +
-                `${requiredScopes.map(scope => `• ${scope}`).join('\n')}\n\n` +
-                `This will allow me to use tools like ${requiresWriteAccess ? 'CreateFile, UpdateFile' : ''} ${requiresExportAccess ? 'ExportData, GenerateReport' : ''}.\n\n` +
-                `Please grant consent to proceed.`,
+        type: "agent",
+        content:
+          `I need your consent to access the following scopes to complete your request:\n\n` +
+          `${requiredScopes.map((scope) => `• ${scope}`).join("\n")}\n\n` +
+          `This will allow me to use tools like ${requiresWriteAccess ? "CreateFile, UpdateFile" : ""} ${requiresExportAccess ? "ExportData, GenerateReport" : ""}.\n\n` +
+          `Please grant consent to proceed.`,
         timestamp: new Date().toISOString(),
-        status: 'completed',
+        status: "completed",
         tools: requiredScopes,
         artifacts: [
           {
-            type: 'url',
-            title: 'Grant Consent',
+            type: "url",
+            title: "Grant Consent",
             content: `/consent-modal?request=${consentRequest.id}`,
-            description: `Grant ${requiredScopes.join(', ')} permissions`
-          }
-        ]
+            description: `Grant ${requiredScopes.join(", ")} permissions`,
+          },
+        ],
       };
 
       chatSessions[sessionId].push(agentMessage);
@@ -143,59 +171,61 @@ export async function action({ request, params }: Route.ActionArgs) {
         success: true,
         requiresConsent: true,
         consentRequestId: consentRequest.id,
-        requiredScopes
+        requiredScopes,
       };
     }
 
     // Add normal agent response
     const agentMessage: Message = {
       id: (Date.now() + 1).toString(),
-      type: 'agent',
+      type: "agent",
       content: `I've successfully processed your request using the available permissions. Here are the results:`,
       timestamp: new Date().toISOString(),
-      status: 'completed',
-      tools: ['ListFiles', 'ReadFile', 'GetFileInfo'],
+      status: "completed",
+      tools: ["ListFiles", "ReadFile", "GetFileInfo"],
       artifacts: [
         {
-          type: 'config',
-          title: 'Task Results',
+          type: "config",
+          title: "Task Results",
           content: `{
   "status": "completed",
   "permissions_used": ["tools:read"],
   "completion_time": "${new Date().toISOString()}",
   "result": "Task completed successfully"
 }`,
-          description: 'Task completed with available permissions'
-        }
-      ]
+          description: "Task completed with available permissions",
+        },
+      ],
     };
 
     chatSessions[sessionId].push(agentMessage);
 
     return {
       success: true,
-      requiresConsent: false
+      requiresConsent: false,
     };
   }
 
-  if (intent === 'approve-consent') {
-    const requestId = formData.get('requestId') as string;
-    const grantedScopes = formData.getAll('scopes') as string[];
+  if (intent === "approve-consent") {
+    const requestId = formData.get("requestId") as string;
+    const grantedScopes = formData.getAll("scopes") as string[];
 
     // Find and update consent request
-    const request = pendingConsentRequests.find(r => r.id === requestId);
+    const request = pendingConsentRequests.find((r) => r.id === requestId);
     if (request) {
-      request.status = 'approved';
+      request.status = "approved";
 
       // Add success message to chat
       if (chatSessions[sessionId]) {
         const successMessage: Message = {
           id: `consent-approved-${Date.now()}`,
-          type: 'agent',
-          content: `✅ Consent granted! I now have access to: ${grantedScopes.join(', ')}\n\nTask completed successfully with your approved permissions.`,
+          type: "agent",
+          content: `✅ Consent granted! I now have access to: ${grantedScopes.join(", ")}\n\nTask completed successfully with your approved permissions.`,
           timestamp: new Date().toISOString(),
-          status: 'completed',
-          tools: grantedScopes.includes('tools:write') ? ['CreateFile', 'UpdateFile'] : ['ExportData', 'GenerateReport']
+          status: "completed",
+          tools: grantedScopes.includes("tools:write")
+            ? ["CreateFile", "UpdateFile"]
+            : ["ExportData", "GenerateReport"],
         };
 
         chatSessions[sessionId].push(successMessage);
@@ -203,12 +233,12 @@ export async function action({ request, params }: Route.ActionArgs) {
 
       return {
         success: true,
-        message: 'Consent approved and task completed',
-        grantedScopes
+        message: "Consent approved and task completed",
+        grantedScopes,
       };
     }
 
-    return { success: false, error: 'Consent request not found' };
+    return { success: false, error: "Consent request not found" };
   }
 
   return null;
@@ -217,23 +247,20 @@ export async function action({ request, params }: Route.ActionArgs) {
 export default function SessionChat({ loaderData }: Route.ComponentProps) {
   const { messages, sessionId, pendingRequests } = loaderData;
   const actionData = useActionData<typeof action>();
-  const [inputMessage, setInputMessage] = useState('');
+  const [inputMessage, setInputMessage] = useState("");
   const messagesEndRef = useRef<HTMLDivElement>(null);
 
   const scrollToBottom = () => {
-    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+    messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
   };
 
   useEffect(() => {
     scrollToBottom();
   }, [messages]);
- 
 
   const copyToClipboard = (text: string) => {
     navigator.clipboard.writeText(text);
   };
-
-  
 
   return (
     <div className="min-h-screen bg-gray-950 text-white">
@@ -256,7 +283,8 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
             <div className="flex items-center space-x-2">
               <Shield className="w-5 h-5 text-yellow-400" />
               <p className="text-yellow-300">
-                Consent required for scopes: {actionData.requiredScopes?.join(', ')}
+                Consent required for scopes:{" "}
+                {actionData.requiredScopes?.join(", ")}
               </p>
             </div>
           </div>
@@ -281,8 +309,12 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
                   <Bot className="w-4 h-4 text-white" />
                 </div>
                 <div>
-                  <h3 className="text-lg font-medium text-white">MCP Consent Assistant</h3>
-                  <p className="text-sm text-gray-400">AI Agent with consent-aware tool access</p>
+                  <h3 className="text-lg font-medium text-white">
+                    MCP Consent Assistant
+                  </h3>
+                  <p className="text-sm text-gray-400">
+                    AI Agent with consent-aware tool access
+                  </p>
                   <p className="text-xs text-blue-400">Session: {sessionId}</p>
                 </div>
               </div>
@@ -306,24 +338,28 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
             {messages.map((message) => (
               <div
                 key={message.id}
-                className={`flex ${message.type === 'user' ? 'justify-end' : 'justify-start'}`}
+                className={`flex ${message.type === "user" ? "justify-end" : "justify-start"}`}
               >
-                <div className={`max-w-xs lg:max-w-md px-4 py-3 rounded-lg ${
-                  message.type === 'user'
-                    ? 'bg-blue-600 text-white'
-                    : message.type === 'system'
-                    ? 'bg-gray-700 text-gray-300'
-                    : 'bg-gray-700 text-white'
-                }`}>
+                <div
+                  className={`max-w-xs lg:max-w-md px-4 py-3 rounded-lg ${
+                    message.type === "user"
+                      ? "bg-blue-600 text-white"
+                      : message.type === "system"
+                        ? "bg-gray-700 text-gray-300"
+                        : "bg-gray-700 text-white"
+                  }`}
+                >
                   <div className="flex items-start space-x-2">
-                    {message.type !== 'user' && (
+                    {message.type !== "user" && (
                       <div className="w-6 h-6 rounded-full bg-gradient-to-r from-green-500 to-blue-500 flex items-center justify-center flex-shrink-0 mt-0.5">
                         <Bot className="w-3 h-3 text-white" />
                       </div>
                     )}
                     <div className="flex-1">
-                      <p className="text-sm whitespace-pre-wrap">{message.content}</p>
-                      
+                      <p className="text-sm whitespace-pre-wrap">
+                        {message.content}
+                      </p>
+
                       {message.tools && (
                         <div className="mt-2 flex flex-wrap gap-1">
                           {message.tools.map((tool, index) => (
@@ -337,44 +373,68 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
                           ))}
                         </div>
                       )}
-                      
-                      {message.status === 'processing' && (
+
+                      {message.status === "processing" && (
                         <div className="mt-2 flex items-center space-x-2">
                           <div className="w-4 h-4 border-2 border-blue-400 border-t-transparent rounded-full animate-spin"></div>
-                          <span className="text-xs text-gray-400">Processing...</span>
+                          <span className="text-xs text-gray-400">
+                            Processing...
+                          </span>
                         </div>
                       )}
-                      
+
                       {/* Artifacts with Consent Modal */}
                       {message.artifacts && message.artifacts.length > 0 && (
                         <div className="mt-4 space-y-3">
-                          <h4 className="text-sm font-medium text-white">Generated Artifacts:</h4>
+                          <h4 className="text-sm font-medium text-white">
+                            Generated Artifacts:
+                          </h4>
                           {message.artifacts.map((artifact, index) => (
-                            <div key={index} className={`p-3 rounded-lg border ${getArtifactColor(artifact.type)}`}>
+                            <div
+                              key={index}
+                              className={`p-3 rounded-lg border ${getArtifactColor(artifact.type)}`}
+                            >
                               <div className="flex items-center justify-between mb-2">
                                 <div className="flex items-center space-x-2">
                                   {getArtifactIcon(artifact.type)}
-                                  <span className="text-sm font-medium">{artifact.title}</span>
+                                  <span className="text-sm font-medium">
+                                    {artifact.title}
+                                  </span>
                                 </div>
                                 <div className="flex items-center space-x-2">
-                                  {artifact.type === 'url' && artifact.content.includes('consent-modal') && (
-                                    <ConsentModal 
-                                      requestId={artifact.content.split('request=')[1]}
-                                      requiredScopes={message.tools || []}
-                                      sessionId={sessionId}
-                                    />
-                                  )}
-                                  {artifact.type === 'url' && !artifact.content.includes('consent-modal') && (
-                                    <button
-                                      onClick={() => window.open(artifact.content, '_blank')}
-                                      className="p-1 hover:bg-white/10 rounded transition-colors duration-200"
-                                      title="Open URL"
-                                    >
-                                      <ExternalLink className="w-3 h-3" />
-                                    </button>
-                                  )}
+                                  {artifact.type === "url" &&
+                                    artifact.content.includes(
+                                      "consent-modal"
+                                    ) && (
+                                      <ConsentModal
+                                        requestId={
+                                          artifact.content.split("request=")[1]
+                                        }
+                                        requiredScopes={message.tools || []}
+                                        sessionId={sessionId}
+                                      />
+                                    )}
+                                  {artifact.type === "url" &&
+                                    !artifact.content.includes(
+                                      "consent-modal"
+                                    ) && (
+                                      <button
+                                        onClick={() =>
+                                          window.open(
+                                            artifact.content,
+                                            "_blank"
+                                          )
+                                        }
+                                        className="p-1 hover:bg-white/10 rounded transition-colors duration-200"
+                                        title="Open URL"
+                                      >
+                                        <ExternalLink className="w-3 h-3" />
+                                      </button>
+                                    )}
                                   <button
-                                    onClick={() => copyToClipboard(artifact.content)}
+                                    onClick={() =>
+                                      copyToClipboard(artifact.content)
+                                    }
                                     className="p-1 hover:bg-white/10 rounded transition-colors duration-200"
                                     title="Copy to clipboard"
                                   >
@@ -383,12 +443,17 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
                                 </div>
                               </div>
                               {artifact.description && (
-                                <p className="text-xs opacity-80 mb-2">{artifact.description}</p>
+                                <p className="text-xs opacity-80 mb-2">
+                                  {artifact.description}
+                                </p>
                               )}
-                              {!artifact.content.includes('consent-modal') && (
+                              {!artifact.content.includes("consent-modal") && (
                                 <div className="bg-black/20 rounded p-2 font-mono text-xs overflow-x-auto">
-                                  {artifact.type === 'code' || artifact.type === 'config' ? (
-                                    <pre className="whitespace-pre-wrap">{artifact.content}</pre>
+                                  {artifact.type === "code" ||
+                                  artifact.type === "config" ? (
+                                    <pre className="whitespace-pre-wrap">
+                                      {artifact.content}
+                                    </pre>
                                   ) : (
                                     <span>{artifact.content}</span>
                                   )}
@@ -427,16 +492,16 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
                 <Send className="w-4 h-4" />
               </button>
             </Form>
-            
+
             <div className="mt-2 flex flex-wrap gap-2">
               {[
-                'Create a new configuration file',
-                'Export my project data',
-                'Can you list my files with current permissions?',
-                'Update my system settings',
-                'Show me my current tool access permissions',
-                'Generate a backup of my data',
-                'Modify the database configuration'
+                "Create a new configuration file",
+                "Export my project data",
+                "Can you list my files with current permissions?",
+                "Update my system settings",
+                "Show me my current tool access permissions",
+                "Generate a backup of my data",
+                "Modify the database configuration",
               ].map((suggestion, index) => (
                 <button
                   key={index}
@@ -455,58 +520,88 @@ export default function SessionChat({ loaderData }: Route.ComponentProps) {
 
   function getToolIcon(tool: string) {
     switch (tool) {
-      case 'ListFiles': return <Database className="w-3 h-3" />;
-      case 'ReadFile': return <Code className="w-3 h-3" />;
-      case 'CreateFile': return <Zap className="w-3 h-3" />;
-      case 'UpdateFile': return <Code className="w-3 h-3" />;
-      case 'DeleteFile': return <Globe className="w-3 h-3" />;
-      case 'ExportData': return <Database className="w-3 h-3" />;
-      case 'GenerateReport': return <Code className="w-3 h-3" />;
-      case 'mcp-guard': return <Globe className="w-3 h-3" />;
-      case 'scope-validator': return <Database className="w-3 h-3" />;
-      case 'consent-manager': return <Zap className="w-3 h-3" />;
-      case 'session-manager': return <MessageCircle className="w-3 h-3" />;
-      case 'token-validator': return <Zap className="w-3 h-3" />;
-      case 'HttpRequest': return <Globe className="w-3 h-3" />;
-      default: return <Zap className="w-3 h-3" />;
+      case "ListFiles":
+        return <Database className="w-3 h-3" />;
+      case "ReadFile":
+        return <Code className="w-3 h-3" />;
+      case "CreateFile":
+        return <Zap className="w-3 h-3" />;
+      case "UpdateFile":
+        return <Code className="w-3 h-3" />;
+      case "DeleteFile":
+        return <Globe className="w-3 h-3" />;
+      case "ExportData":
+        return <Database className="w-3 h-3" />;
+      case "GenerateReport":
+        return <Code className="w-3 h-3" />;
+      case "mcp-guard":
+        return <Globe className="w-3 h-3" />;
+      case "scope-validator":
+        return <Database className="w-3 h-3" />;
+      case "consent-manager":
+        return <Zap className="w-3 h-3" />;
+      case "session-manager":
+        return <MessageCircle className="w-3 h-3" />;
+      case "token-validator":
+        return <Zap className="w-3 h-3" />;
+      case "HttpRequest":
+        return <Globe className="w-3 h-3" />;
+      default:
+        return <Zap className="w-3 h-3" />;
     }
   }
 
- 
-
   function getArtifactIcon(type: string) {
     switch (type) {
-      case 'url': return <ExternalLink className="w-4 h-4" />;
-      case 'config': return <Code className="w-4 h-4" />;
-      case 'code': return <Code className="w-4 h-4" />;
-      case 'documentation': return <Database className="w-4 h-4" />;
-      default: return <Code className="w-4 h-4" />;
+      case "url":
+        return <ExternalLink className="w-4 h-4" />;
+      case "config":
+        return <Code className="w-4 h-4" />;
+      case "code":
+        return <Code className="w-4 h-4" />;
+      case "documentation":
+        return <Database className="w-4 h-4" />;
+      default:
+        return <Code className="w-4 h-4" />;
     }
   }
 
   function getArtifactColor(type: string) {
     switch (type) {
-      case 'url': return 'bg-blue-500/20 text-blue-400 border-blue-500/30';
-      case 'config': return 'bg-green-500/20 text-green-400 border-green-500/30';
-      case 'code': return 'bg-purple-500/20 text-purple-400 border-purple-500/30';
-      case 'documentation': return 'bg-orange-500/20 text-orange-400 border-orange-500/30';
-      default: return 'bg-gray-500/20 text-gray-400 border-gray-500/30';
+      case "url":
+        return "bg-blue-500/20 text-blue-400 border-blue-500/30";
+      case "config":
+        return "bg-green-500/20 text-green-400 border-green-500/30";
+      case "code":
+        return "bg-purple-500/20 text-purple-400 border-purple-500/30";
+      case "documentation":
+        return "bg-orange-500/20 text-orange-400 border-orange-500/30";
+      default:
+        return "bg-gray-500/20 text-gray-400 border-gray-500/30";
     }
   }
 }
 
 // Consent Modal Component
-function ConsentModal({ requestId, requiredScopes, sessionId }: { requestId: string; requiredScopes: string[]; sessionId: string }) {
+function ConsentModal({
+  requestId,
+  requiredScopes,
+  sessionId,
+}: {
+  requestId: string;
+  requiredScopes: string[];
+  sessionId: string;
+}) {
   const [isOpen, setIsOpen] = useState(false);
   const consentFetcher = useFetcher();
 
   const handleApprove = () => {
     const formData = new FormData();
-    formData.set('intent', 'approve-consent');
-    formData.set('requestId', requestId);
-    requiredScopes.forEach(scope => formData.append('scopes', scope));
-    
-    consentFetcher.submit(formData, { method: 'post' });
+    formData.set("intent", "approve-consent");
+    formData.set("requestId", requestId);
+    requiredScopes.forEach((scope) => formData.append("scopes", scope));
+
+    consentFetcher.submit(formData, { method: "post" });
     setIsOpen(false);
   };
 
@@ -531,14 +626,19 @@ function ConsentModal({ requestId, requiredScopes, sessionId }: { requestId: str
               <Shield className="w-6 h-6 text-blue-400" />
               <h3 className="text-lg font-bold text-white">Consent Required</h3>
             </div>
-            
+
             <p className="text-gray-300 mb-4">
-              The agent is requesting access to the following permissions for session <span className="font-mono text-blue-400">{sessionId}</span>:
+              The agent is requesting access to the following permissions for
+              session{" "}
+              <span className="font-mono text-blue-400">{sessionId}</span>:
             </p>
-            
+
             <div className="space-y-2 mb-6">
               {requiredScopes.map((scope, index) => (
-                <div key={index} className="flex items-center space-x-2 p-2 bg-blue-500/10 rounded border border-blue-500/20">
+                <div
+                  key={index}
+                  className="flex items-center space-x-2 p-2 bg-blue-500/10 rounded border border-blue-500/20"
+                >
                   <Shield className="w-4 h-4 text-blue-400" />
                   <span className="text-sm text-blue-300">{scope}</span>
                 </div>
@@ -547,8 +647,9 @@ function ConsentModal({ requestId, requiredScopes, sessionId }: { requestId: str
 
             <div className="bg-yellow-500/10 border border-yellow-500/20 rounded-lg p-3 mb-6">
               <p className="text-xs text-yellow-300">
-                By granting consent, you allow the agent to use the specified tools to complete your request.
-                This permission will be active for this session only and can be revoked at any time.
+                By granting consent, you allow the agent to use the specified
+                tools to complete your request. This permission will be active
+                for this session only and can be revoked at any time.
               </p>
             </div>
 
@@ -561,10 +662,12 @@ function ConsentModal({ requestId, requiredScopes, sessionId }: { requestId: str
               </button>
               <button
                 onClick={handleApprove}
-                disabled={consentFetcher.state !== 'idle'}
+                disabled={consentFetcher.state !== "idle"}
                 className="px-4 py-2 bg-green-600 hover:bg-green-700 text-white rounded-lg transition-colors disabled:opacity-50"
               >
-                {consentFetcher.state !== 'idle' ? 'Granting...' : 'Grant Consent'}
+                {consentFetcher.state !== "idle"
+                  ? "Granting..."
+                  : "Grant Consent"}
               </button>
             </div>
