--- app/portal/deno.jsonc
@@ -1,50 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/denoland/deno/refs/heads/main/cli/schemas/config-file.v1.json",
-  "nodeModulesDir": "auto",
-  "unstable": [
-    "sloppy-imports"
-  ],
-  "tasks": {
-    "build": "deno run -A --unstable-detect-cjs npm:@react-router/dev@7.6.1 build",
-    "dev": "deno run -A --unstable-detect-cjs npm:@react-router/dev@7.6.1 dev",
-    "start": "deno run --allow-env --allow-net --allow-read ./server.ts",
-    "typegen": "deno run -A --unstable-detect-cjs npm:@react-router/dev@7.6.1 typegen",
-    "typecheck": {
-      "command": "deno check .",
-      "dependencies": [
-        "typegen"
-      ]
-    }
-  },
-  "imports": {
-    "@std/http": "jsr:@std/http@^1.0.16",
-    "~/": "./app/",
-    "@deno/vite-plugin": "npm:@deno/vite-plugin@1.0.4",
-    "@react-router/dev": "npm:@react-router/dev@7.6.1",
-    "@react-router/serve": "npm:@react-router/serve@7.6.1",
-    "@std/assert": "jsr:@std/assert@1",
-    "@tailwindcss/vite": "npm:@tailwindcss/vite@4.1.13",
-    "@types/react-dom": "npm:@types/react-dom@19.1.5",
-    "@types/react": "npm:@types/react@19.1.5",
-    "isbot": "npm:isbot@5.1.28",
-    "react-dom": "npm:react-dom@19.1.0",
-    "react-router": "npm:react-router@7.6.1",
-    "react": "npm:react@19.1.0",
-    "tailwindcss": "npm:tailwindcss@4.1.13",
-    "vite": "npm:vite@6.3.5",
-    "lucide-react": "npm:lucide-react@0.544.0"
-  },
-  "compilerOptions": {
-    "verbatimModuleSyntax": true,
-    "strict": true,
-    "jsx": "react-jsx",
-    "jsxImportSource": "react",
-    "types": [
-      "vite/client"
-    ],
-    "rootDirs": [
-      ".",
-      "./.react-router/types"
-    ]
-  }
-}
