+++ app/portal/package.json
@@ -1,8 +1,38 @@
 {
-  "comment": "This only exists to help vite module resolution. Do not use this file, instread use deno.json.",
+  "name": "u-p",
+  "private": true,
+  "type": "module",
   "scripts": {
-    "build": "deno task build",
-    "start": "deno task start",
-    "dev": "deno task dev"
+    "build": "react-router build",
+    "dev": "cross-env NODE_ENV=development node server.js",
+    "start": "node server.js",
+    "typecheck": "react-router typegen && tsc -b"
+  },
+  "dependencies": {
+    "@react-router/express": "^7.7.1",
+    "@react-router/node": "^7.7.1",
+    "compression": "^1.7.5",
+    "express": "^5.1.0",
+    "isbot": "^5.1.27",
+    "morgan": "^1.10.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^7.7.1"
+  },
+  "devDependencies": {
+    "@react-router/dev": "^7.7.1",
+    "@tailwindcss/vite": "^4.1.4",
+    "@types/compression": "^1.7.5",
+    "@types/express": "^5.0.1",
+    "@types/express-serve-static-core": "^5.0.6",
+    "@types/morgan": "^1.9.9",
+    "@types/node": "^20",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "cross-env": "^7.0.3",
+    "tailwindcss": "^4.1.4",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.3",
+    "vite-tsconfig-paths": "^5.1.4"
   }
 }
