+++ app/portal/vite.config.ts
@@ -1,20 +1,15 @@
 import { reactRouter } from "@react-router/dev/vite";
-import deno from "@deno/vite-plugin";
 import tailwindcss from "@tailwindcss/vite";
 import { defineConfig } from "vite";
-// import tsconfigPaths from "vite-tsconfig-paths";
+import tsconfigPaths from "vite-tsconfig-paths";
 
-export default defineConfig({
-  plugins: [ tailwindcss(),reactRouter(), deno(),],
-  environments: {
-    ssr: {
-      build: {
-        target: "ESNext",
-      },
-      resolve: {
-        conditions: ["deno"],
-        externalConditions: ["deno"],
-      },
-    },
+export default defineConfig(({ isSsrBuild }) => ({
+  build: {
+    rollupOptions: isSsrBuild
+      ? {
+          input: "./server/app.ts",
+        }
+      : undefined,
   },
-});
+  plugins: [tailwindcss(), reactRouter(), tsconfigPaths()],
+}));
