--- app/user-portal/app/app.css
@@ -1,291 +1,0 @@
-@import "tailwindcss";
-
-/* Custom styles for Agent Grants platform */
-
-/* Dark theme optimizations */
-:root {
-  --bg-primary: #0f172a;
-  --bg-secondary: #1e293b;
-  --bg-tertiary: #334155;
-  --text-primary: #f8fafc;
-  --text-secondary: #cbd5e1;
-  --text-tertiary: #94a3b8;
-  --accent-blue: #3b82f6;
-  --accent-green: #10b981;
-  --accent-red: #ef4444;
-  --accent-yellow: #f59e0b;
-  --accent-purple: #8b5cf6;
-}
-
-/* Smooth transitions for all interactive elements */
-* {
-  transition:
-    color 0.2s ease,
-    background-color 0.2s ease,
-    border-color 0.2s ease,
-    transform 0.2s ease;
-}
-
-/* Custom scrollbar for dark theme */
-::-webkit-scrollbar {
-  width: 8px;
-  height: 8px;
-}
-
-::-webkit-scrollbar-track {
-  background: var(--bg-secondary);
-  border-radius: 4px;
-}
-
-::-webkit-scrollbar-thumb {
-  background: var(--bg-tertiary);
-  border-radius: 4px;
-}
-
-::-webkit-scrollbar-thumb:hover {
-  background: #475569;
-}
-
-/* Focus styles for accessibility */
-button:focus-visible,
-input:focus-visible,
-select:focus-visible,
-textarea:focus-visible,
-a:focus-visible {
-  outline: 2px solid var(--accent-blue);
-  outline-offset: 2px;
-}
-
-/* Loading animations */
-@keyframes spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.5;
-  }
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.animate-spin {
-  animation: spin 1s linear infinite;
-}
-
-.animate-pulse {
-  animation: pulse 2s cubic-bezier(0.4, 0, 0.6, 1) infinite;
-}
-
-.animate-fade-in {
-  animation: fadeIn 0.3s ease-out;
-}
-
-/* Glass morphism effects */
-.glass {
-  background: rgba(255, 255, 255, 0.05);
-  -webkit-backdrop-filter: blur(10px);
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.glass-dark {
-  background: rgba(0, 0, 0, 0.2);
-  -webkit-backdrop-filter: blur(10px);
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.05);
-}
-
-/* Custom button styles */
-.btn-primary {
-  @apply bg-blue-600 hover:bg-blue-700 text-white px-4 py-2 rounded-lg font-medium transition-colors duration-200 focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 focus:ring-offset-gray-900;
-}
-
-.btn-secondary {
-  @apply bg-gray-600 hover:bg-gray-700 text-white px-4 py-2 rounded-lg font-medium transition-colors duration-200 focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 focus:ring-offset-gray-900;
-}
-
-.btn-success {
-  @apply bg-green-600 hover:bg-green-700 text-white px-4 py-2 rounded-lg font-medium transition-colors duration-200 focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:ring-offset-gray-900;
-}
-
-.btn-danger {
-  @apply bg-red-600 hover:bg-red-700 text-white px-4 py-2 rounded-lg font-medium transition-colors duration-200 focus:ring-2 focus:ring-red-500 focus:ring-offset-2 focus:ring-offset-gray-900;
-}
-
-/* Status indicators */
-.status-online {
-  @apply bg-green-500 text-green-500;
-}
-
-.status-offline {
-  @apply bg-red-500 text-red-500;
-}
-
-.status-warning {
-  @apply bg-yellow-500 text-yellow-500;
-}
-
-.status-info {
-  @apply bg-blue-500 text-blue-500;
-}
-
-/* Card styles */
-.card {
-  @apply bg-gray-800/50 backdrop-blur-sm rounded-xl border border-gray-700 p-6;
-}
-
-.card-header {
-  @apply border-b border-gray-600 pb-4 mb-4;
-}
-
-.card-footer {
-  @apply border-t border-gray-600 pt-4 mt-4;
-}
-
-/* Form styles */
-.form-input {
-  @apply w-full bg-gray-700 border border-gray-600 rounded-lg px-3 py-2 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500 focus:ring-1 focus:ring-blue-500;
-}
-
-.form-select {
-  @apply w-full bg-gray-700 border border-gray-600 rounded-lg px-3 py-2 text-white focus:outline-none focus:border-blue-500 focus:ring-1 focus:ring-blue-500;
-}
-
-.form-textarea {
-  @apply w-full bg-gray-700 border border-gray-600 rounded-lg px-3 py-2 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500 focus:ring-1 focus:ring-blue-500 resize-none;
-}
-
-/* Badge styles */
-.badge {
-  @apply inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium;
-}
-
-.badge-success {
-  @apply bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400;
-}
-
-.badge-warning {
-  @apply bg-yellow-100 text-yellow-800 dark:bg-yellow-900/20 dark:text-yellow-400;
-}
-
-.badge-error {
-  @apply bg-red-100 text-red-800 dark:bg-red-900/20 dark:text-red-400;
-}
-
-.badge-info {
-  @apply bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-400;
-}
-
-.badge-gray {
-  @apply bg-gray-100 text-gray-800 dark:bg-gray-900/20 dark:text-gray-400;
-}
-
-/* Tooltip styles */
-.tooltip {
-  @apply absolute z-10 px-2 py-1 text-xs text-white bg-gray-900 rounded shadow-lg opacity-0 pointer-events-none transition-opacity duration-200;
-}
-
-.tooltip.show {
-  @apply opacity-100;
-}
-
-/* Modal styles */
-.modal-overlay {
-  @apply fixed inset-0 bg-black/50 flex items-center justify-center z-50;
-}
-
-.modal-content {
-  @apply bg-gray-800 rounded-xl border border-gray-700 max-w-2xl w-full mx-4 max-h-[90vh] overflow-y-auto;
-}
-
-/* Chat message styles */
-.chat-message {
-  @apply max-w-xs lg:max-w-md px-4 py-3 rounded-lg;
-}
-
-.chat-message.user {
-  @apply bg-blue-600 text-white ml-auto;
-}
-
-.chat-message.agent {
-  @apply bg-gray-700 text-white;
-}
-
-.chat-message.system {
-  @apply bg-gray-700 text-gray-300;
-}
-
-/* Grant card styles */
-.grant-card {
-  @apply bg-gray-700/50 rounded-lg p-4 border border-gray-600 hover:bg-gray-700/70 transition-colors duration-200;
-}
-
-.grant-card.granted {
-  @apply border-green-500/30;
-}
-
-.grant-card.denied {
-  @apply border-red-500/30;
-}
-
-/* Audit entry styles */
-.audit-entry {
-  @apply p-4 hover:bg-gray-700/30 transition-colors duration-200 border-l-4;
-}
-
-.audit-entry.success {
-  @apply border-green-500;
-}
-
-.audit-entry.warning {
-  @apply border-yellow-500;
-}
-
-.audit-entry.error {
-  @apply border-red-500;
-}
-
-.audit-entry.info {
-  @apply border-blue-500;
-}
-
-/* Responsive utilities */
-@media (max-width: 640px) {
-  .mobile-hidden {
-    display: none;
-  }
-
-  .mobile-full {
-    width: 100%;
-  }
-}
-
-/* Print styles */
-@media print {
-  .no-print {
-    display: none !important;
-  }
-
-  .print-only {
-    display: block !important;
-  }
-}
